Move sync_codec into behaviour
